requirements.txt
streamlit
langchain
langchain-community
python-dotenv
GitPython
langchain_openai
chromadb
